Use int64 type to extract `int` attributes in TF-TensorRT
TensorFlow always uses int64 to store `int` attributes.
https://github.com/tensorflow/tensorflow/blob/60089e430e34c870e074c6d5746ed88ea9b39909/tensorflow/core/framework/attr_value.proto#L32

PiperOrigin-RevId: 235666747

+++ tensorflow/compiler/tf2tensorrt/convert/convert_nodes_test.cc
@@ -2553,8 +2553,8 @@ TEST_F(OpConverterTest, ConvertStridedSlice) {
 
   // Get nodedef for StridedSlice layer.
   auto get_strided_slice_nodedef =
-      [](int begin_mask = 0, int end_mask = 0, int ellipsis_mask = 0,
-         int new_axis_mask = 0, int shrink_axis_mask = 0) -> NodeDef {
+      [](int64 begin_mask = 0, int64 end_mask = 0, int64 ellipsis_mask = 0,
+         int64 new_axis_mask = 0, int64 shrink_axis_mask = 0) -> NodeDef {
     Scope s = Scope::NewRootScope();
     auto input = ops::Placeholder(s.WithOpName("input"), DT_FLOAT);
     auto begin = ops::Placeholder(s.WithOpName("begin"), DT_INT32);
